🚧 Components typing folder
