+++ src/types/components/AnimalCard/index.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from 'react';
+
+export interface IAnimalCardProps {
+  title: string;
+  imageUrl: string | null;
+  headerOptions?: ReactNode;
+  body?: ReactNode;
+  headerBackground: string;
+  onPress?: () => void;
+}
