+++ src/types/components/CustomTextInput/index.ts
@@ -1,0 +1,7 @@
+import { FormikProps } from 'formik';
+
+export interface ICustomTextInput<T> {
+  formikHelpers: FormikProps<T>,
+  fieldName: keyof T,
+  [propName: string]: unknown;
+}
