+++ src/types/components/DrawerContent/index.ts
@@ -1,0 +1,4 @@
+export interface IDrawerProps {
+  parentDrawerOpen: boolean,
+  setParentDrawerOpen: (_: boolean) => void,
+}
